\name{cellauto-package}
\alias{cellauto-package}
\alias{cellauto}
\docType{package}
\title{
\packageTitle{cellauto}
}
\description{
\packageDescription{cellauto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cellauto}
\packageIndices{cellauto}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{cellauto}

Maintainer: \packageMaintainer{cellauto}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
